feat: feature layouts(common) added
